requirements.txt
dvc
jupyter
matplotlib
numpy
pandas
psutil
PyYAML
scikit-learn
catboost